Skip to end if updating answer
