+++ core/templates/core/summary/section-affected-goods.html
@@ -2,7 +2,7 @@
 
 <div class="govuk-summary-list__row">
   <dt class="govuk-summary-list__key">What goods are affected?</dt>
-  <dd class="govuk-summary-list__value">{{ summary.commodity.commodity_code|join:'' }} - {{ summary.commodity.label }}</dd>
+  <dd class="govuk-summary-list__value">{{ summary.commodity.commodity_code|join:'' }} - {{ summary.commodity.label|safe }}</dd>
   <dd class="govuk-summary-list__actions">
     <a class="govuk-link" href="{% get_step_url 'product-search' %}">Change<span class="govuk-visually-hidden"> affected your answer</span></a>
   </dd>
