+++ core/templates/core/wizard-step-product.html
@@ -41,8 +41,8 @@
           </div>
           <p class="govuk-body">                
             {% for item in result.hierarchy_context %}
-              {% if item.0.description|striptags %}
-                <span class="govuk-link govuk-!-font-size-16 {% if forloop.last %}breadcrumb-item-last{%  else %}breadcrumb-item {% endif %}">{{ item.0.description }}</span>
+              {% if item.0.description %}
+                <span class="govuk-link govuk-!-font-size-16 {% if forloop.last %}breadcrumb-item-last{%  else %}breadcrumb-item {% endif %}">{{ item.0.description|striptags }}</span>
               {% endif %}
             {% endfor %}
           </p>
